[pyassimp] Use numpy when possible
It allows for faster access to data, brings useful methods and is
pyOpenGL-friendly

It eases manipulation of matrices and vectors (many operations like
product, transposition, etc. are available)

numpy.array are also directly usable in many places in pyopengl.

And make sure we store floats as float32 to please OpenGL

Update sample.py to be happy with numpy arrays