[pyassimp] Support Python3
This include the renaming of __init__.py in core.py to solve tricky imports issues

sample.py runs correctly, except that the ctype structure returned is not
complete. The issue is very similar to the one we have with 32bits/64bits
issues

Removed the PyAssimp3 subproject project.

To maintain source compatibility between py2 and py3, python >= 2.6 is now required