DefaultIOStream::FileSize(): Replaced windows-based version to use mFile directly instead of fileStat
Unfortunately, it is not 100% clear anymore what's the background of this change. Maybe it was faster or it avoided issues in case that file loaders used exclusive open.
All I know is that this change is not the accidental result of a merge, because according to SVN blame, this code section had never been touched since revision 1. So it must have been done for InfraWorks for some reason.
So I preserved this change to avoid regression due to the AssImp update.