ObjImporter: Removed unnecessary removal of '\' chars.
std::vector::erase() has complexity O(n) which for a large buffer is very
slow.

Additionally in ObjFileParser we have a copyNextLine() function which
takes line continuation with backslashes into account. This is used only
for faces ATM. I'm not sure when a '\' can occur in a .obj file, but I've
only ever seen it on face lines that have more than 3 vertices.

We could use copyNextLine() for other types of lines too if needed. This
would slow down parsing of lines a little bit, but that is probably
mitigated by not scanning through removing backslashes first (that took
1/3rd of the import time in my file that has no backslashes).

Finally I've updated the progress reporting to take into account this
change.