Solved pyassimp.errors.AssimpError in conda
This error occurred because of error in finding .extension files.
 
os.path.splitext(filename)[-1].lower() not in ext_whitelist:  ### this line had bugs, failed to select files with extensions .so.x  [.so.1  .so.2  .so.3.1] but worked on simple extensions only. like- .so .dll

Added a mechanism to remove this bug.