Implemented full pivot support without fake transforms
Pivot support was previously quite hard to maintain so we opted to improve and rewrite it the way that should work the best for assimp.

This method uses the standard pivot formula and apply's this to all node transforms as per the specifications.

This implements many standard things we were missing in FBX too relating to FBXPose, cluster handling and fixes a lot of bugs regarding importing pivots.

Previously we could not import pivots properly, now we can. This will need cleaned up and is a proof of concept for the 'production ready version'

we have known issues but we are working on them.

In future re-sampling can be removed as not required all of the time. 
todo: add code to interpolate keyframes and apply pivot xform data.