Removed uneeded expression in else() and endif() constructs.
The following changes were tested against building with a master copy.

For a more detailed report, please read the followings below:

The flags that were checked were the ones checked by default + the following:

ASSIMP_BUILD_ASSIMP_VIEW
ASSIMP_BUILD_SAMPLES

The list of compilable projects in Visual Studio were the following:

- assimp
- assimp_cmd
- assimp_simpleogl
- assimp_simpletextureddirectx11
- assimp_simpletexturedogl
- assimp_viewer
- INSTALL
- IrrXML
- uninstall
- unit
- UpdateAssimpLibsDebugSymbolsAndDLLs
- ZERO_CHECK
- zlib
- zlibstatic

The solution that was generated with the new changes comprised the same number of projects as the one from the master, unmodified. All projects except UpdateAssimpLibsDebugSymbolsAndDLLs were built without any error. All tests were run.

UpdateAssimpLibsDebugSymbolsAndDLLs could not be built because of the reasons explained in issue #3004.