Raised minimum CMake version to 3.0 for assimp_cmd and assimp_view.
This is a follow-up to PR #3024, in which I explained that I would then change the version of CMake to 3.0 in several files to match the changes that were originally made by OP in PR #3008.

The minimum CMake version have been raised to 3.0 in both assimp_cmd and assimp_view projects.

The same CMake configuration as in PR #3024 was used. All projects from PR #3024 are found in the solution with the current changes. All projects built. All tests were run.