Fix dynamic loading path for OSX
This commit sets the INSTALL_RPATH to $ORIGIN on not OSX, or the
OSX-sort-of-equivalent @loader_path. It tells the dynamic loader to
check the path 'from the binary directory to the library directory'.
This fixes not being able to find IrrXML.dylib when things are not
installed to a default location. This fixes issue
https://github.com/assimp/assimp/issues/2872 for me.

The patch is pretty much straight from the book 'professional cmake'.