X3D exporter: Modified meta data comment in exported files
At first I was going to fix the typo ("tou" -> "you"), but decided to just drop the details out of the comment completely.

A comment about MetadataString and Scene nodes is very specific to assimp, and wouldn't really mean anything to anybody else looking at the file. Also it's for developers, and is neither meaningful nor useful to users. So, I removed the details.

Notes about exporter behavior should go in assimp's documentation, where it is meaningful and also a lot easier to find. It shouldn't go in exported files.