contrib/unzip is now a submodule
Assimp’s original minizip version cannot be found on the internet any more. When switching to submodules, it had to be replaced with version 1.2, which is similar, but has subtle differences in the encryption implementation. This introduced the error “undefined reference to `arc4random_buf'” on BSD.

To mitigate, I disabled minizip encryption support by defining “NOCRYPT”.